IntroCompanyController: 기존 게임 소개 매핑 컨트롤러에서 회사 소개글 컨트롤러 분리  기존 파일처리 클래스(FileUtil) 관련 메소드를 FileUploadUtil 클래스의 메소드로 대체
