+++ src/main/resources/templates/detail/company-list.html
@@ -29,23 +29,6 @@
       </ul>
     </div>
   </div>
-<!--  <div class="menu-allList">-->
-<!--    <div id="gameList">-->
-<!--      <div th:each="intro : ${companyList}" id="gList">-->
-<!--        <img th:src="@{/upload/{upload}(upload=${intro.upload})}"-->
-<!--             width="260px" height="180px">-->
-<!--        <ul>-->
-<!--          <li><a th:href="@{getIntro(id=${intro.id})}" style="cursor: pointer" th:text="${intro.name}"></a>-->
-<!--          </li>-->
-<!--          <li th:text="${intro.id}"></li>-->
-<!--          <li><img th:src="@{/images/img-icon-rPc.png}"></li>-->
-<!--        </ul>-->
-<!--      </div>-->
-<!--    </div>-->
-<!--  </div>-->
-
-
-<!--  <div class="Frame2">-->
     <div class="Group18" th:each="intro:${introList}">
       <div class="Group17">
         <div class="Rectangle39"></div>
@@ -77,15 +60,16 @@
   /*<![CDATA[*/
   const element = document.getElementById('board-top-menu');
   const loginMember = /*[[${session.loginMember}]]*/ '';
-  if(loginMember.memberId != null || loginMember === 'null'){
+  let status = /*[[${session.status}]]*/'';
+  if((loginMember != null && loginMember.memberId != null) || loginMember === null){
     element.style.visibility = 'hidden';
-  }else if(loginMember.companyId != null  && '${status}' === 'writeOnce'){
+  }else if(loginMember.companyId != null  && status === 'writeOnce'){
     $('button[name="listWrite"]').on("click", function() {
       alert("글 작성은 한번만 가능합니다!");
     });
   }else if(loginMember.companyId != null){
     $('button[name="listWrite"]').on("click", function() {
-      location.href = "../detail/insertIntro";
+      location.href = "../insertIntro";
     });
   }
   /*]]>*/
