+++ src/main/resources/templates/detail/company-regi.html
@@ -25,8 +25,7 @@
       <p>Game_registration</p>
       <div id="clearFix"></div>
     </div>
-    <form  action="/detail/insertIntro" method="post" enctype="multipart/form-data">
-      <input style="visibility: hidden" type="text" name="cId" class="default-write" th:value="${company?.companyId}">
+    <form  action="/insertIntro" method="post" enctype="multipart/form-data">
       <table>
         <tr>
           <th>회사 명</th>
