+++ src/main/resources/templates/main/header.html
@@ -84,7 +84,7 @@
         <li><a href="/alllist">All List</a></li>
         <li><a href="/#menu-main1">Recommendations</a></li>
         <li><a href="/notice/list">Notice</a></li>
-        <li><a href="/detail/getIntroList">Team Intro</a></li>
+        <li><a href="/getIntroList">Team Intro</a></li>
       </ul>
     </nav>
   </div>
